+++ fractals/src/main/java/me/catmousedog/fractals/canvas/Mouse.java
@@ -6,8 +6,8 @@ import java.awt.event.MouseListener;
 import javax.swing.SwingUtilities;
 
 import me.catmousedog.fractals.fractals.LinearTransform;
+import me.catmousedog.fractals.ui.Configuration;
 import me.catmousedog.fractals.ui.JPInterface;
-import me.catmousedog.fractals.ui.Logger;
 
 /**
  * the mouse listener for interacting with the canvas
@@ -20,9 +20,9 @@ public class Mouse implements MouseListener {
 	private final Canvas canvas;
 
 	/**
-	 * the logger instance
+	 * the configuration of the canvas
 	 */
-	private final Logger logger;
+	private final Configuration config;
 
 	/**
 	 * the user interface, used to {@link JPInterface#render()} and retrieve user
@@ -30,35 +30,30 @@ public class Mouse implements MouseListener {
 	 */
 	private JPInterface jpi;
 
-	private double zoomFactor = 1.5;
-
-	public Mouse(Canvas canvas, Logger logger) {
+	public Mouse(Canvas canvas) {
 		this.canvas = canvas;
-		this.logger = logger;
+		this.config = canvas.getConfig();
 	}
 
 	@Override
 	public void mouseReleased(MouseEvent me) {
 		me = SwingUtilities.convertMouseEvent(me.getComponent(), me, canvas);
 
-		LinearTransform transform = canvas.getTransform();
+		// JPInterface#renderWithout(Runnable r) couldn't be used because of the
+		// MouseEvent here
+		jpi.save();
+
+		LinearTransform transform = config.getTransform();
 		double[] t = transform.apply(me.getX(), me.getY());
 		transform.setTranslation(t[0], t[1]);
 
-		// manually save zoomFactor as it is needed for the transformation
-		try {
-			zoomFactor = Double.parseDouble(jpi.getZoomJTF().getText());
-		} catch (NumberFormatException e) {
-			logger.log("zoom factor must be a valid double");
-		}
-
 		// lmb (zoom in)
 		if (me.getButton() == MouseEvent.BUTTON1)
-			transform.zoom(1 / zoomFactor);
+			transform.zoom(1 / config.getZoomFactor());
 
 		// rmb (zoom out)
 		if (me.getButton() == MouseEvent.BUTTON3)
-			transform.zoom(zoomFactor);
+			transform.zoom(config.getZoomFactor());
 
 		jpi.update();
 		jpi.renderNow();