+++ fractals/src/main/java/me/catmousedog/fractals/components/Data.java
@@ -2,14 +2,14 @@ package me.catmousedog.fractals.components;
 
 import java.awt.event.ActionListener;
 
-import javax.swing.JComponent;
-
 /**
- * A data container in the form of one or multiple {@link JComponent}s.<br>
- * An instance of {@link Data} can be 'active', meaning it takes in a
- * {@link ActionListener} as a parameter to the constructor i.e. it actively
- * performs an action. Or it can be 'passive', meaning the data has to be
- * retrieved manually through the {@link Data#getData()} when needed.
+ * A data container extending the {@link Item} class. This class just represents
+ * an {@link Item} that holds data.<br>
+ * <p>
+ * <b>Active</b> {@link Item}s have an {@link ActionListener} and perform an
+ * action themselves.
+ * <p>
+ * <b>Passive</b> {@link Item}s don't perform an action.
  * 
  * @param <T> The type of the stored data. <br>
  *            For example: a {@link Double} for a textfield.
@@ -18,17 +18,33 @@ public abstract class Data<T> extends Item {
 
 	protected T data;
 
+	/**
+	 * gets the data without saving
+	 * 
+	 * @return the current {@link Data#data}
+	 */
 	public T getData() {
 		return data;
 	}
 
+	/**
+	 * saves and gets the data
+	 * 
+	 * @return calls {@link Data#save()} and then returns {@link Data#data}
+	 */
+	public T saveAndGet() {
+		save();
+		return data;
+	}
+
+	/**
+	 * sets the data
+	 * 
+	 * @param data sets the {@link Data#data} field and calls {@link Data#update()}
+	 */
 	public void setData(T data) {
 		this.data = data;
 		update();
 	}
 
-	public T saveAndGet() {
-		save();
-		return data;
-	}
 }