+++ fractals/src/main/java/me/catmousedog/fractals/components/Padding.java
@@ -1,0 +1,29 @@
+package me.catmousedog.fractals.components;
+
+import java.awt.Component;
+import java.awt.Dimension;
+
+import javax.swing.Box;
+
+public class Padding extends Item {
+
+	private int h;
+	
+	public Padding(int h) {
+		this.h = h;
+	}
+	
+	@Override
+	public Component panel() {
+		return Box.createRigidArea(new Dimension(0, h));
+	}
+
+	@Override
+	public void save() {
+	}
+
+	@Override
+	public void update() {
+	}
+
+}